@@ -14,29 +14,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.4",
-    "@angular/common": "7.2.4",
-    "@angular/compiler": "7.2.4",
-    "@angular/core": "7.2.4",
-    "@angular/forms": "7.2.4",
-    "@angular/platform-browser": "7.2.4",
-    "@angular/platform-browser-dynamic": "7.2.4",
-    "@angular/router": "7.2.4",
-    "core-js": "^2.4.1",
+    "@angular/animations": "~13.3.0",
+    "@angular/common": "~13.3.0",
+    "@angular/compiler": "~13.3.0",
+    "@angular/core": "~13.3.0",
+    "@angular/forms": "~13.3.0",
+    "@angular/platform-browser": "~13.3.0",
+    "@angular/platform-browser-dynamic": "~13.3.0",
+    "@angular/router": "~13.3.0",
+    "rxjs": "~7.5.0",
     "marked": "^0.3.9",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.29"
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.1",
-    "@angular/compiler-cli": "7.2.4",
-    "@angular/language-service": "7.2.4",
-    "@testing-library/cypress": "^8.0.1",
-    "@types/node": "^9.4.0",
-    "codelyzer": "^4.5.0",
-    "cypress": "^8.5.0",
-    "tslint": "~5.9.1"
+    "@angular-devkit/build-angular": "~13.3.0",
+    "@angular/cli": "~13.3.0",
+    "@angular/compiler-cli": "~13.3.0",
+    "@types/jasmine": "~3.10.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~4.0.0",
+    "typescript": "~4.6.2"
   }
 }
