@@ -2,39 +2,40 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
 import { User, UserService, Profile } from '../core';
-import { concatMap ,  tap } from 'rxjs/operators';
+import { concatMap, tap } from 'rxjs/operators';
 
 @Component({
   selector: 'app-profile-page',
-  templateUrl: './profile.component.html'
+  templateUrl: './profile.component.html',
 })
 export class ProfileComponent implements OnInit {
   constructor(
     private route: ActivatedRoute,
     private userService: UserService
-  ) { }
+  ) {}
 
   profile: Profile;
   currentUser: User;
   isUser: boolean;
 
   ngOnInit() {
-    this.route.data.pipe(
-      concatMap((data: { profile: Profile }) => {
-        this.profile = data.profile;
-        // Load the current user's data.
-        return this.userService.currentUser.pipe(tap(
-          (userData: User) => {
-            this.currentUser = userData;
-            this.isUser = (this.currentUser.username === this.profile.username);
-          }
-        ));
-      })
-    ).subscribe();
+    this.route.data
+      .pipe(
+        concatMap((data: any) => {
+          this.profile = data.profile;
+          // Load the current user's data.
+          return this.userService.currentUser.pipe(
+            tap((userData: User) => {
+              this.currentUser = userData;
+              this.isUser = this.currentUser.username === this.profile.username;
+            })
+          );
+        })
+      )
+      .subscribe();
   }
 
   onToggleFollowing(following: boolean) {
     this.profile.following = following;
   }
-
 }
