@@ -3,11 +3,11 @@ import { Router } from '@angular/router';
 
 import { Article, ArticlesService, UserService } from '../../core';
 import { of } from 'rxjs';
-import { concatMap ,  tap } from 'rxjs/operators';
+import { concatMap, tap } from 'rxjs/operators';
 
 @Component({
   selector: 'app-favorite-button',
-  templateUrl: './favorite-button.component.html'
+  templateUrl: './favorite-button.component.html',
 })
 export class FavoriteButtonComponent {
   constructor(
@@ -16,45 +16,48 @@ export class FavoriteButtonComponent {
     private userService: UserService
   ) {}
 
-  @Input() article: Article;
+  @Input() article: Article = null;
   @Output() toggle = new EventEmitter<boolean>();
   isSubmitting = false;
 
   toggleFavorite() {
     this.isSubmitting = true;
 
-    this.userService.isAuthenticated.pipe(concatMap(
-      (authenticated) => {
-        // Not authenticated? Push to login screen
-        if (!authenticated) {
-          this.router.navigateByUrl('/login');
-          return of(null);
-        }
+    this.userService.isAuthenticated
+      .pipe(
+        concatMap((authenticated) => {
+          // Not authenticated? Push to login screen
+          if (!authenticated) {
+            this.router.navigateByUrl('/login');
+            return of(null);
+          }
 
-        // Favorite the article if it isn't favorited yet
-        if (!this.article.favorited) {
-          return this.articlesService.favorite(this.article.slug)
-          .pipe(tap(
-            data => {
-              this.isSubmitting = false;
-              this.toggle.emit(true);
-            },
-            err => this.isSubmitting = false
-          ));
+          // Favorite the article if it isn't favorited yet
+          if (!this.article.favorited) {
+            return this.articlesService.favorite(this.article.slug).pipe(
+              tap(
+                (data) => {
+                  this.isSubmitting = false;
+                  this.toggle.emit(true);
+                },
+                (err) => (this.isSubmitting = false)
+              )
+            );
 
-        // Otherwise, unfavorite the article
-        } else {
-          return this.articlesService.unfavorite(this.article.slug)
-          .pipe(tap(
-            data => {
-              this.isSubmitting = false;
-              this.toggle.emit(false);
-            },
-            err => this.isSubmitting = false
-          ));
-        }
-
-      }
-    )).subscribe();
+            // Otherwise, unfavorite the article
+          } else {
+            return this.articlesService.unfavorite(this.article.slug).pipe(
+              tap(
+                (data) => {
+                  this.isSubmitting = false;
+                  this.toggle.emit(false);
+                },
+                (err) => (this.isSubmitting = false)
+              )
+            );
+          }
+        })
+      )
+      .subscribe();
   }
 }
