@@ -4,29 +4,35 @@ import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
 const routes: Routes = [
   {
     path: 'settings',
-    loadChildren: './settings/settings.module#SettingsModule'
+    loadChildren: () =>
+      import('./settings/settings.module').then((x) => x.SettingsModule),
   },
   {
     path: 'profile',
-    loadChildren: './profile/profile.module#ProfileModule'
+    loadChildren: () =>
+      import('./profile/profile.module').then((x) => x.ProfileModule),
   },
   {
     path: 'editor',
-    loadChildren: './editor/editor.module#EditorModule'
+    loadChildren: () =>
+      import('./editor/editor.module').then((x) => x.EditorModule),
   },
   {
     path: 'article',
-    loadChildren: './article/article.module#ArticleModule'
-  }
+    loadChildren: () =>
+      import('./article/article.module').then((x) => x.ArticleModule),
+  },
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes, {
-    // preload all modules; optionally we could
-    // implement a custom preloading strategy for just some
-    // of the modules (PRs welcome 😉)
-    preloadingStrategy: PreloadAllModules
-  })],
-  exports: [RouterModule]
+  imports: [
+    RouterModule.forRoot(routes, {
+      // preload all modules; optionally we could
+      // implement a custom preloading strategy for just some
+      // of the modules (PRs welcome 😉)
+      preloadingStrategy: PreloadAllModules,
+    }),
+  ],
+  exports: [RouterModule],
 })
 export class AppRoutingModule {}
