@@ -8,12 +8,12 @@ import {
   Comment,
   CommentsService,
   User,
-  UserService
+  UserService,
 } from '../core';
 
 @Component({
   selector: 'app-article-page',
-  templateUrl: './article.component.html'
+  templateUrl: './article.component.html',
 })
 export class ArticleComponent implements OnInit {
   article: Article;
@@ -21,7 +21,7 @@ export class ArticleComponent implements OnInit {
   canModify: boolean;
   comments: Comment[];
   commentControl = new FormControl();
-  commentFormErrors = {};
+  commentFormErrors: any = {};
   isSubmitting = false;
   isDeleting = false;
 
@@ -30,28 +30,25 @@ export class ArticleComponent implements OnInit {
     private articlesService: ArticlesService,
     private commentsService: CommentsService,
     private router: Router,
-    private userService: UserService,
-  ) { }
+    private userService: UserService
+  ) {}
 
   ngOnInit() {
     // Retreive the prefetched article
-    this.route.data.subscribe(
-      (data: { article: Article }) => {
-        this.article = data.article;
+    this.route.data.subscribe((data: any) => {
+      this.article = data.article;
 
-        // Load the comments on this article
-        this.populateComments();
-      }
-    );
+      // Load the comments on this article
+      this.populateComments();
+    });
 
     // Load the current user's data
-    this.userService.currentUser.subscribe(
-      (userData: User) => {
-        this.currentUser = userData;
+    this.userService.currentUser.subscribe((userData: User) => {
+      this.currentUser = userData;
 
-        this.canModify = (this.currentUser.username === this.article.author.username);
-      }
-    );
+      this.canModify =
+        this.currentUser.username === this.article.author.username;
+    });
   }
 
   onToggleFavorite(favorited: boolean) {
@@ -71,17 +68,15 @@ export class ArticleComponent implements OnInit {
   deleteArticle() {
     this.isDeleting = true;
 
-    this.articlesService.destroy(this.article.slug)
-      .subscribe(
-        success => {
-          this.router.navigateByUrl('/');
-        }
-      );
+    this.articlesService.destroy(this.article.slug).subscribe((success) => {
+      this.router.navigateByUrl('/');
+    });
   }
 
   populateComments() {
-    this.commentsService.getAll(this.article.slug)
-      .subscribe(comments => this.comments = comments);
+    this.commentsService
+      .getAll(this.article.slug)
+      .subscribe((comments) => (this.comments = comments));
   }
 
   addComment() {
@@ -89,28 +84,24 @@ export class ArticleComponent implements OnInit {
     this.commentFormErrors = {};
 
     const commentBody = this.commentControl.value;
-    this.commentsService
-      .add(this.article.slug, commentBody)
-      .subscribe(
-        comment => {
-          this.comments.unshift(comment);
-          this.commentControl.reset('');
-          this.isSubmitting = false;
-        },
-        errors => {
-          this.isSubmitting = false;
-          this.commentFormErrors = errors;
-        }
-      );
+    this.commentsService.add(this.article.slug, commentBody).subscribe(
+      (comment) => {
+        this.comments.unshift(comment);
+        this.commentControl.reset('');
+        this.isSubmitting = false;
+      },
+      (errors) => {
+        this.isSubmitting = false;
+        this.commentFormErrors = errors;
+      }
+    );
   }
 
   onDeleteComment(comment) {
-    this.commentsService.destroy(comment.id, this.article.slug)
-      .subscribe(
-        success => {
-          this.comments = this.comments.filter((item) => item !== comment);
-        }
-      );
+    this.commentsService
+      .destroy(comment.id, this.article.slug)
+      .subscribe((success) => {
+        this.comments = this.comments.filter((item) => item !== comment);
+      });
   }
-
 }
