@@ -5,31 +5,25 @@ import { ArticleListConfig, Profile } from '../core';
 
 @Component({
   selector: 'app-profile-articles',
-  templateUrl: './profile-articles.component.html'
+  templateUrl: './profile-articles.component.html',
 })
 export class ProfileArticlesComponent implements OnInit {
-  constructor(
-    private route: ActivatedRoute,
-    private router: Router
-  ) {}
+  constructor(private route: ActivatedRoute, private router: Router) {}
 
   profile: Profile;
   articlesConfig: ArticleListConfig = {
     type: 'all',
-    filters: {}
+    filters: {},
   };
 
   ngOnInit() {
-    this.route.parent.data.subscribe(
-      (data: {profile: Profile}) => {
-        this.profile = data.profile;
-        this.articlesConfig = {
-          type: 'all',
-          filters: {}
-        }; // Only method I found to refresh article load on swap
-        this.articlesConfig.filters.author = this.profile.username;
-      }
-    );
+    this.route.parent.data.subscribe((data: any) => {
+      this.profile = data.profile;
+      this.articlesConfig = {
+        type: 'all',
+        filters: {},
+      }; // Only method I found to refresh article load on swap
+      this.articlesConfig.filters.author = this.profile.username;
+    });
   }
-
 }
