@@ -5,27 +5,21 @@ import { ArticleListConfig, Profile } from '../core';
 
 @Component({
   selector: 'app-profile-favorites',
-  templateUrl: './profile-favorites.component.html'
+  templateUrl: './profile-favorites.component.html',
 })
 export class ProfileFavoritesComponent implements OnInit {
-  constructor(
-    private route: ActivatedRoute,
-    private router: Router
-  ) {}
+  constructor(private route: ActivatedRoute, private router: Router) {}
 
   profile: Profile;
   favoritesConfig: ArticleListConfig = {
     type: 'all',
-    filters: {}
+    filters: {},
   };
 
   ngOnInit() {
-    this.route.parent.data.subscribe(
-      (data: {profile: Profile}) => {
-        this.profile = data.profile;
-        this.favoritesConfig.filters.favorited = this.profile.username;
-      }
-    );
+    this.route.parent.data.subscribe((data: any) => {
+      this.profile = data.profile;
+      this.favoritesConfig.filters.favorited = this.profile.username;
+    });
   }
-
 }
